no need to typecast to int pointers for calculating the object index

+++ src/slab.c
@@ -69,11 +69,11 @@ int free_object(slab_desc* slab_desc_p, void* object, cache* cachep)
 		return 0;
 
 	// object must be a multiple of object_size away from the first object in the slab
-	if(((((uintptr_t)object) - ((uintptr_t)slab_desc_p->objects)) % cachep->object_size))
+	if((object - slab_desc_p->objects) % cachep->object_size)
 		return 0;
 
 	// calculate the index of the object
-	unsigned int object_index = (((uintptr_t)object) - ((uintptr_t)slab_desc_p->objects)) / cachep->object_size;
+	unsigned int object_index = (object - slab_desc_p->objects) / cachep->object_size;
 
 	// this effectively lets us know that there is a free object close by the last allocated index
 	if(slab_desc_p->free_objects == 0)
